refactor: 불필요한 레디스 레포지토리 삭제 및 폴더 정리
- Lua 스크립트를 사용함에 따라 불필요해진 레포지토리와 테스트 코드 삭제
- purchase 폴더 안에 있던 Ticket 관련 레포지토리를 ticket 폴더 안으로 이동
