--- backend/core/build.gradle
@@ -1,2 +1,0 @@
-dependencies {
-}
